Renamed fixture, const and context files and updated imports
