+++ src/modules/pos/validator/__fixtures__/mockPosConstants.js
