+++ src/modules/pos/validator/consts/validators.js
