+++ src/modules/pos/validator/context/validatorRowContext.js
